Refactor: 라우트, 컨트롤러 및 서비스 모듈을 도메인에 맞춰 통합
