+++ src/controllers/bugReport/index.ts
@@ -2,8 +2,18 @@ import { Request, Response } from 'express';
 import { sendError } from '@/utils/response';
 import { AuthenticatedRequest } from '@/types/express';
 import { StatusCodes } from 'http-status-codes';
-import { getAllBugReports, fetchPostDetail } from '@/services/bugListService';
-import { GetBugReportListRequestBody,GetBugReportsResponse,GetBugReportDetailsResponse} from '@/dto/reportDto';
+import {
+  GetBugReportListRequestBody,
+  GetBugReportsResponse,
+  GetBugReportDetailsResponse,
+  CreateBugReportRequestBody, CreateBugReportResponse, CreateBugImageResponse,
+} from '@/dto/reportDto';
+import { uploadToS3 } from '@/utils/upload';
+import {
+  createBugReport,
+  fetchPostDetail,
+  getAllBugReports,
+} from '@/services/bugReport';
 
 // 사냥 리스트 조회
 export const getBugReportList = async (req: Request<GetBugReportListRequestBody>, res: Response<GetBugReportsResponse>)=>{
@@ -18,14 +28,14 @@ export const getBugReportList = async (req: Request<GetBugReportListRequestBody>
 
     res.status(StatusCodes.OK).json(bugReports);
   } catch (error) {
-    console.error('Error fetching bug reports:', error);
+    console.error('Error fetching bugReport reports:', error);
     return sendError(res, '버그 리포트를 가져오는 중 오류가 발생했습니다..');
   }
 };
 
 // 사냥 상세 정보 조회
 export const getBugReportDetail = async (
-  req: AuthenticatedRequest<{id: string},{}>, 
+  req: AuthenticatedRequest<{id: string},{}>,
   res: Response<GetBugReportDetailsResponse>
 ) => {
   if(! req.user) {
@@ -36,7 +46,7 @@ export const getBugReportDetail = async (
     const { id }  = req.params;
     console.log(Number(id))
     if (!id || isNaN(Number(id))) {
-      return sendError(res, '유효한 버그 리포트 ID를 제공해야 합니다.'); 
+      return sendError(res, '유효한 버그 리포트 ID를 제공해야 합니다.');
     }
 
     const bugReportDetail = await fetchPostDetail(Number(id));
@@ -46,7 +56,38 @@ export const getBugReportDetail = async (
 
     res.status(StatusCodes.OK).json(bugReportDetail);
   } catch (error) {
-    console.error('Error fetching bug report details:', error);
+    console.error('Error fetching bugReport report details:', error);
     return sendError(res, '버그 리포트 상세 정보를 가져오는 중 오류가 발생했습니다.',500);
   }
 };
+
+// 사냥 게시글 작성
+export const createBugPost = async (req: AuthenticatedRequest<{}, CreateBugReportRequestBody>,
+                                    res: Response<CreateBugReportResponse>) => {
+  if(! req.user) {
+    return sendError(res, '로그인된 사용자가 아닙니다.', StatusCodes.UNAUTHORIZED);
+  }
+  try {
+    await createBugReport(req.body, req.user);
+    res.status(StatusCodes.CREATED).send();
+  } catch(e) {
+    console.error(e);
+    return sendError(res, '게시글 업로드에 실패했습니다.');
+  }
+};
+
+// 벌레 사진 입력
+export const uploadBugImage = async (req: Request, res: Response<CreateBugImageResponse>) => {
+  if(! req.file) {
+    return sendError(res, '파일이 업로드되지 않았습니다.');
+  }
+  try {
+    const url = await uploadToS3(req.file);
+    res.json({
+      image_url: url,
+    });
+  } catch(e) {
+    console.error(e);
+    return sendError(res, '파일 업로드에 실패했습니다.');
+  }
+}
