+++ src/index.ts
@@ -1,8 +1,6 @@
 import express, { NextFunction, Request, Response } from 'express';
 import 'dotenv/config';
 import { kakaoLoginRoute } from './routes/kakaoLoginRoute';
-import { bugPostRoute } from './routes/bugPostRoute';
-import { bugListRoute } from './routes/bugListRoute';
 import { userRoute } from '@/routes/user';
 import cors from 'cors';
 import fs from 'node:fs';
@@ -11,6 +9,7 @@ import * as path from 'node:path';
 import { sendError } from '@/utils/response';
 import { StatusCodes } from 'http-status-codes';
 import { matchRoute } from '@/routes/match';
+import { bugReportRoute } from '@/routes/bugReport';
 
 const app = express();
 const DEV_PORT = 3000;
@@ -31,8 +30,7 @@ app.use((req: Request, res: Response, next: NextFunction) => {
 });
 
 app.use(API_PREFIX, kakaoLoginRoute);
-app.use(API_PREFIX, bugPostRoute);
-app.use(API_PREFIX, bugListRoute);
+app.use(API_PREFIX, bugReportRoute);
 app.use(API_PREFIX, userRoute);
 app.use(API_PREFIX, matchRoute);
